+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/_utils.py
@@ -7,7 +7,7 @@ from knack.util import CLIError
 from knack.log import get_logger
 from azure.cli.core.commands.parameters import get_resources_in_subscription
 
-from azure.mgmt.containerregistry.v2017_10_01.models import SkuName, Sku
+from azure.mgmt.containerregistry.v2018_02_01_preview.models import SkuName, Sku
 
 from ._constants import (
     REGISTRY_RESOURCE_TYPE,
@@ -39,7 +39,7 @@ def _arm_get_resource_by_name(cli_ctx, resource_name, resource_type):
         message = "The resource with name '{}' and type '{}' could not be found".format(
             resource_name, resource_type)
         try:
-            subscription = profile.get_subscription()
+            subscription = profile.get_subscription(cli_ctx.data['subscription_id'])
             raise CLIError("{} in subscription '{} ({})'.".format(message, subscription['name'], subscription['id']))
         except (KeyError, TypeError):
             raise CLIError("{} in the current subscription.".format(message))
@@ -238,18 +238,20 @@ def _parameters(registry_name,
     return parameters
 
 
-# pylint: disable=inconsistent-return-statements
 def random_storage_account_name(cli_ctx, registry_name):
     from datetime import datetime
 
     client = get_storage_service_client(cli_ctx).storage_accounts
     prefix = registry_name[:18].lower()
 
-    while True:
+    for x in range(10):
         time_stamp_suffix = datetime.utcnow().strftime('%H%M%S')
         storage_account_name = ''.join([prefix, time_stamp_suffix])[:24]
+        logger.debug("Checking storage account %s with name '%s'.", x, storage_account_name)
         if client.check_name_availability(storage_account_name).name_available:  # pylint: disable=no-member
             return storage_account_name
+
+    raise CLIError("Could not find an available storage account name. Please try again later.")
 
 
 def validate_managed_registry(cli_ctx, registry_name, resource_group_name=None, message=None):