+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/credential.py
@@ -8,17 +8,15 @@ from knack.util import CLIError
 from ._utils import get_registry_by_name
 
 
-# pylint: disable=inconsistent-return-statements
 def acr_credential_show(cmd, client, registry_name, resource_group_name=None):
     registry, resource_group_name = get_registry_by_name(cmd.cli_ctx, registry_name, resource_group_name)
 
     if registry.admin_user_enabled:  # pylint: disable=no-member
         return client.list_credentials(resource_group_name, registry_name)
 
-    admin_not_enabled_error(registry_name)
+    raise admin_not_enabled_error(registry_name)
 
 
-# pylint: disable=inconsistent-return-statements
 def acr_credential_renew(cmd, client, registry_name, password_name, resource_group_name=None):
     registry, resource_group_name = get_registry_by_name(cmd.cli_ctx, registry_name, resource_group_name)
 
@@ -26,9 +24,8 @@ def acr_credential_renew(cmd, client, registry_name, password_name, resource_gro
         return client.regenerate_credential(
             resource_group_name, registry_name, password_name)
 
-    admin_not_enabled_error(registry_name)
+    raise admin_not_enabled_error(registry_name)
 
 
 def admin_not_enabled_error(registry_name):
-    raise CLIError("Run 'az acr update -n {} --admin-enabled true' to enable admin first.".format(
-        registry_name))
+    return CLIError("Run 'az acr update -n {} --admin-enabled true' to enable admin first.".format(registry_name))