+++ src/azure-cli-core/HISTORY.rst
@@ -2,9 +2,14 @@
 
 Release History
 ===============
-2.0.39
+
+2.0.40
 ++++++
 * authentication: support authorization code flow for interactive login
+
+2.0.39
+++++++
+* MSI packaging change
 
 2.0.38
 ++++++