+++ src/command_modules/azure-cli-acr/HISTORY.rst
@@ -6,6 +6,8 @@ Release History
 2.0.28
 ++++++
 * Add polling build status.
+* Allow case-insensitive enum values.
+* Add --top and --orderby parameters for show-manifests.
 
 2.0.27
 ++++++